feat(auth): added confirm password page. Closes #669
